refactor(region_map): move persistence logic to region_map_persistence.js and update runtime imports
